[Perf] CUDA + AMDGPU launchers: pin stream to null at launcher scope; drop redundant pre/post-kernel stream_synchronize
