+++ quadrants/runtime/llvm/adstack_lazy_claim/metadata_publish.cpp
@@ -31,7 +31,7 @@
 #include "quadrants/rhi/amdgpu/amdgpu_context.h"
 #endif
 
-#include "quadrants/runtime/llvm/adstack_lazy_claim/cuda_stream_pin.h"
+#include "quadrants/rhi/cuda/cuda_stream_pin.h"
 
 namespace quadrants::lang {
